backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Clerk authentication
pyjwt[crypto]>=2.8.0
httpx>=0.25.0

# PostgreSQL driver - only needed for production
# psycopg2-binary==2.9.10

